+++ platformio.ini
@@ -13,8 +13,6 @@ env_default = Master
 src_dir     = car
 
 [env:Dashboard]
-lib_deps =
-    SDFileSystem
 platform = ststm32
 board = nucleo_f446re
 framework = mbed
@@ -30,8 +28,6 @@ build_unflags =
     -std=gnu++98
 
 [env:Pedal]
-lib_deps =
-    SDFileSystem
 platform = ststm32
 board = nucleo_f446re
 framework = mbed
@@ -47,8 +43,6 @@ build_unflags =
     -std=gnu++98
 
 [env:Master]
-lib_deps =
-    SDFileSystem
 platform = ststm32
 board = nucleo_f767zi
 framework = mbed
@@ -67,7 +61,6 @@ build_unflags =
 lib_compat_mode = off
 lib_ignore =
     ArduinoSTL
-    SDFileSystem
 lib_ldf_mode = chain+
 platform = teensy
 board = teensy31
@@ -83,8 +76,6 @@ build_unflags =
     -std=gnu++98
 
 [env:Dashboard_new_small]
-lib_deps =
-    SDFileSystem
 platform = ststm32
 board = nucleo_l432kc
 framework = mbed
@@ -101,8 +92,6 @@ build_unflags =
     -std=gnu++98
 
 [env:Pedal_new_small]
-lib_deps =
-    SDFileSystem
 platform = ststm32
 board = nucleo_l432kc
 framework = mbed
@@ -147,8 +136,6 @@ build_unflags =
     -std=gnu++98
 
 [env:F767ZI_TESTING]
-lib_deps =
-    SDFileSystem
 platform = ststm32
 board = nucleo_f767zi
 framework = mbed
@@ -182,7 +169,6 @@ lib_deps =
 lib_ignore =
     carpi
     ArduinoSTL
-    SDFileSystem
 lib_ldf_mode = chain+
 platform = teensy
 board = teensy31
